// src/app/shared/models/models.ts
export interface apiResultFormat {
  data: [];
  totalData: number;
}
// ...
export interface HotelDetails {
  isSelected?: boolean;
  sNo?: number;
  id: string; 
  img:string;
  hotel: string; 
  room: string; 
  guest: string; 
  days: string; 
  location:string;
  pricing: string; 
  bookedOn: string; 
  status: string; 
}